third_party/blink/web_tests/editing/execCommand/queryCommandState-04.html

<!doctype HTML>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div contenteditable="true">
<sub><i>1. Make text subscript and italic</i></sub>
<sup><i>2. Make text superscript and italic</i></sup>
</div>
<div id="log"></div>
<script>
function test_sub() {
    document.getSelection().removeAllRanges();
    var range = document.createRange();
    range.selectNode(document.querySelector('i'));
    document.getSelection().addRange(range);

    assert_equals(document.queryCommandState('subscript'), true);

    document.getSelection().removeAllRanges();
    assert_false(document.queryCommandState('subscript'));

    // Test command state after collapsing the selection to the end of the first
    // <i> element.
    const textNode = document.querySelector('i').firstChild;
    document.getSelection().collapse(textNode, textNode.textContent.length);

    assert_true(document.queryCommandState('subscript'));
}

function test_sup() {
    document.getSelection().removeAllRanges();
    var range = document.createRange();
    range.selectNode(document.querySelectorAll('i')[1]);
    document.getSelection().addRange(range);

    assert_equals(document.queryCommandState('superscript'), true);

    document.getSelection().removeAllRanges();
    assert_false(document.queryCommandState('superscript'));

    // Test command state after collapsing the selection to the end of the second
    // <i> element.
    const textNode = document.querySelectorAll('i')[1].firstChild;
    document.getSelection().collapse(textNode, textNode.textContent.length);

    assert_true(document.queryCommandState('superscript'));
}

function test_all(platform) {
    if (platform)
        internals.settings.setEditingBehavior(platform);
    test(test_sub, `${platform}: run queryCommandState('subscript')`);
    test(test_sup, `${platform}: run queryCommandState('superscript')`);
}

(window.internals ? ['mac', 'win',''] : ['']).forEach(test_all);
</script>
